Preliminarily completed app relinquish cpu

+++ kernel/string.c
@@ -29,7 +29,7 @@ int32 memcmp(const void *src1, const void *src2, uint64 len) {
     return res;
 }
 
-void memmove(uint8 *dest, const uint8 *src, uint32 len) {
+void memmove(uint8 *dest, const uint8 *src, uint64 len) {
     uint8* temp = dest;
     for ( ; len > 0; len--) {
         *temp++ = *src++;
